Module Tezos_protocol_006_PsCARTHA__Protocol.Delegate_storage

type balance =

| Contract of Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t

| Rewards of Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t * Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t

| Fees of Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t * Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t

| Deposits of Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t * Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t

Places where tezzies can be found in the ledger's state.

type balance_update =

| Debited of Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t

| Credited of Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t

A credit or debit of tezzies to a balance.

type balance_updates = (balance * balance_update) list

A list of balance updates. Duplicates may happen.

val balance_updates_encoding : balance_updates Tezos_protocol_environment_006_PsCARTHA__Environment.Data_encoding.t

val cleanup_balance_updates : balance_updates -> balance_updates

Remove zero-valued balances from a list of updates.

type frozen_balance = {

deposit : Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t;

fees : Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t;

rewards : Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t;

}

val init : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Allow to register a delegate when creating an account.

val remove : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Cleanup delegation when deleting a contract.

val get : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t option Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Reading the current delegate of a contract.

val registered : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val set : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t option -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Updating the delegate of a contract.

When calling this function on an "implicit contract" and setting the delegate to the contract manager registers it as a delegate. One cannot unregister a delegate for now. The associate contract is now 'undeletable'.

type Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.error += No_deletion of Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t | Active_delegate | Current_delegate | Empty_delegate_account of Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t | Balance_too_low_for_deposit of {

delegate : Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t;

deposit : Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t;

balance : Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t;

}

val fold : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> init:'a -> f:(Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> 'a -> 'a Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t) -> 'a Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Iterate on all registered delegates.

val list : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t list Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

List all registered delegates.

val freeze_deposit : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Various functions to 'freeze' tokens. A frozen 'deposit' keeps its associated rolls. When frozen, 'fees' may trigger new rolls allocation. Rewards won't trigger new rolls allocation until unfrozen.

val freeze_fees : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val freeze_rewards : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t -> Tezos_raw_protocol_006_PsCARTHA.Raw_context.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val cycle_end : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t -> Tezos_raw_protocol_006_PsCARTHA.Nonce_storage.unrevealed list -> (Tezos_raw_protocol_006_PsCARTHA.Raw_context.t * balance_updates * Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t list) Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Trigger the context maintenance at the end of cycle 'n', i.e.: unfreeze deposit/fees/rewards from 'n - preserved_cycle' ; punish the provided unrevealed seeds (tipically seed from cycle 'n - 1'). Returns a list of account with the amount that was unfrozen for each and the list of deactivated delegates.

val punish : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t -> (Tezos_raw_protocol_006_PsCARTHA.Raw_context.t * frozen_balance) Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Burn all then frozen deposit/fees/rewards for a delegate at a given cycle. Returns the burned amounts.

val has_frozen_balance : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t -> bool Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Has the given key some frozen tokens in its implicit contract?

val frozen_balance : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Returns the amount of frozen deposit, fees and rewards associated to a given delegate.

val frozen_balance_encoding : frozen_balance Tezos_protocol_environment_006_PsCARTHA__Environment.Data_encoding.t

val frozen_balance_by_cycle_encoding : frozen_balance Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.Map.t Tezos_protocol_environment_006_PsCARTHA__Environment.Data_encoding.t

val frozen_balance_by_cycle : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> frozen_balance Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.Map.t Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Returns the amount of frozen deposit, fees and rewards associated to a given delegate, indexed by the cycle by which at the end the balance will be unfrozen.

val full_balance : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Returns the full 'balance' of the implicit contract associated to a given key, i.e. the sum of the spendable balance and of the frozen balance.

val staking_balance : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val delegated_contracts : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Contract_repr.t list Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

Returns the list of contracts (implicit or originated) that delegated towards a given delegate

val delegated_balance : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Tez_repr.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val deactivated : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t

val grace_period : Tezos_raw_protocol_006_PsCARTHA.Raw_context.t -> Tezos_protocol_environment_006_PsCARTHA__Environment.Signature.Public_key_hash.t -> Tezos_raw_protocol_006_PsCARTHA.Cycle_repr.t Tezos_protocol_environment_006_PsCARTHA__Environment.Error_monad.tzresult Tezos_protocol_environment_006_PsCARTHA__Environment.Lwt.t