Module Tezos_raw_protocol_006_PsCARTHA.Level_repr

type t = private {

level : Raw_level_repr.t;
(*
The level of the block relative to genesis. This is also the Shell's notion of level.
*)

level_position : int32;
(*
The level of the block relative to the block that starts protocol alpha. This is specific to the protocol alpha. Other protocols might or might not include a similar notion.
*)

cycle : Cycle_repr.t;
(*
The current cycle's number. Note that cycles are a protocol-specific notion. As a result, the cycle number starts at 0 with the first block of protocol alpha.
*)

cycle_position : int32;
(*
The current level of the block relative to the first block of the current cycle.
*)

voting_period : Voting_period_repr.t;

voting_period_position : int32;

expected_commitment : bool;

}

type level = t

include Tezos_protocol_environment_006_PsCARTHA.Compare.S with type t := level

val (=) : level -> level -> bool

val (<>) : level -> level -> bool

val (<) : level -> level -> bool

val (<=) : level -> level -> bool

val (>=) : level -> level -> bool

val (>) : level -> level -> bool

val compare : level -> level -> int

val equal : level -> level -> bool

val max : level -> level -> level

val min : level -> level -> level

val encoding : level Tezos_protocol_environment_006_PsCARTHA.Data_encoding.t

val pp : Tezos_protocol_environment_006_PsCARTHA.Format.formatter -> level -> unit

val pp_full : Tezos_protocol_environment_006_PsCARTHA.Format.formatter -> level -> unit

val root : Raw_level_repr.t -> level

val from_raw : first_level:Raw_level_repr.t -> blocks_per_cycle:int32 -> blocks_per_voting_period:int32 -> blocks_per_commitment:int32 -> Raw_level_repr.t -> level

val diff : level -> level -> int32