Module Tezos_raw_protocol_006_PsCARTHA__Storage

Tezos Protocol Implementation - Typed storage

This module hides the hierarchical (key x value) database under pre-allocated typed accessors for all persistent entities of the tezos context.

This interface enforces no invariant on the contents of the database. Its goal is to centralize all accessors in order to have a complete view over the database contents and avoid key collisions.